BUG FIX - Instant Task Dropdown Positioning (#79)
Modifies the positioning of the dropdown shown by the "Instant Task" button on a Job Screen. Originally, it was positioned off the left side of the button (possibly overflowing from the screen). This change positions it aligned to the right side of the button, with the options being shown under the other buttons in this section.